Fix buttet and collision bug. Bullet now fires one at a time. Zombie dies when health goes below or equals to zero

+++ Assets/Scripts/ZombieAttack.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class ZombieAttack : MonoBehaviour {
-	private int distance = 2;
+	private float distance = 2f;
 
 	private Animator anim;
 	private GameObject player;
